+++ wms-client/src/components/BO/Cloth/ModifyStock/TypeExchangeBoard.js
@@ -94,14 +94,31 @@ class TypeExchangeBoard extends Component {
   handleSubmitClick() {
     const { oldClothInfo } = this.state;
     let newClothInfoes = this.state.newClothInfoes;
+    let oldTotalLength = parseFloat(oldClothInfo.clothIdentifier.length);
+    let totalLength;
     for (let i = 0; i < newClothInfoes.length; i += 1) {
       delete newClothInfoes[i]["errors"];
+      totalLength += parseFloat(newClothInfoes[i].length);
     }
-    this.props.purgeOldClothInfoNotExist(oldClothInfo.clothIdentifier.id);
-    this.props.typeExchangeBatchCreateClothInfo(
-      newClothInfoes,
-      this.props.history
-    );
+    if (oldTotalLength === totalLength) {
+      if (window.confirm("確認送出？")) {
+        this.props.purgeOldClothInfoNotExist(oldClothInfo.clothIdentifier.id);
+        this.props.typeExchangeBatchCreateClothInfo(
+          newClothInfoes,
+          this.props.history
+        );
+      }
+    } else {
+      // call backend create file
+      // if user still confirm to process, auto send file to user
+      if (window.confirm("減肥前後總長度不符，確認送出？")) {
+        this.props.purgeOldClothInfoNotExist(oldClothInfo.clothIdentifier.id);
+        this.props.typeExchangeBatchCreateClothInfo(
+          newClothInfoes,
+          this.props.history
+        );
+      }
+    }
   }
 
   render() {
@@ -118,7 +135,7 @@ class TypeExchangeBoard extends Component {
               >
                 返回
               </button>
-              <p className="h3 text-center">異動資料</p>
+              <p className="h3 text-center">異動前狀態</p>
               <hr />
               <table className="table">
                 <thead>
